Disable fromClusterClaim when namespace restrictions are set
In commit 612affb, an attempt was made to restrict fromClusterClaim to a
specific namespace when the namespace is restricted (LookupNamespace).
The issue is that ClusterClaim is a cluster resource and not a namespace
resource so this approach did not work.

This commit makes it so that fromClusterClaim raises an error when the
namespace is restricted (LookupNamespace).

Signed-off-by: mprahl <mprahl@users.noreply.github.com>