Return an error when a lookup references an unknown API resource (#38)
The `ResolveTemplate` method will still function the same way except
that an error is also returned along with the resolved template when a
`lookup` template function call references an unknown API resource (e.g.
a CRD isn't installed). This lets the caller optionally choose to
refresh the `KubeAPIResourceList` value and try to resolve the templates
again.

Relates:
https://github.com/stolostron/backlog/issues/20382

Signed-off-by: mprahl <mprahl@users.noreply.github.com>