Migrate from glog to klog
klog is a drop-in replacement of glog and is more compatible with
container deployments and the rest of the Open Cluster Management Policy
stack.

This also removes the default log level constant since that really
should be a level 0 log and each log message should have a level
explicitly chosen.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>