Restrict lookups of cluster-scoped resources
In some cases, it was possible to do a cluster-scoped lookup, despite
`LookupNamespace` being set on the config, which was intended to prevent
this. That is fixed, and more thoroughly tested.

A new config field `ClusterScopedAllowList` was added to allow specific
cluster-scoped resources to be used in `lookup` when `LookupNamespace`
is set.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>