Add two caching modes
The default caching mode is one that caches objects for the duration of
the ResolveTemplate execution. This mode is not completely concurrency
safe.

The advanced caching mode is concurrency safe and utilizes watches to
update the cache in background goroutines.

The resolve options are now passed to the ResolveTemplate method. This
allows a shared resolver to resolve templates for different
propagated policies. This allows the cache to be shared as well.

Relates:
https://issues.redhat.com/browse/ACM-7398

Signed-off-by: mprahl <mprahl@users.noreply.github.com>