Loosen template context restrictions
The allowed types for the template context (i.e. variables) are structs,
map[string]string, strings, and a recursive combination of the above.

This is to allow a template variable that contains a subset of the
Kubernetes metadata field (i.e. a map with string and map[string]string
values).

Relates:
https://issues.redhat.com/browse/ACM-12566

Signed-off-by: mprahl <mprahl@users.noreply.github.com>