Give a better error for unwatchable resources
Previously, templates that lookup resources which do not support the
`watch` API could have various errors, especially if caching was used.
The dependency-watcher library now has an error specifically for this
situation, which we can pass on to users. Generally the solution is to
use a non-caching resolver for those resources, which should still work.

Note: the test for this behavior needed a resource type that supports
`get` but not `watch`. The deprecated ComponentStatus works for now, but
might be removed eventually.

Refs:
 - https://issues.redhat.com/browse/ACM-19965

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>