Add `lookupClusterClaim`
This new function has basically the same functionality as using the
`lookup` function, in particular it does not stop the templating with an
error if the specified ClusterClaim does not exist. This is especially
helpful for ClusterClaims (in comparison to the analogous functions for
ConfigMaps and Secrets) because the apiVersion for ClusterClaims is
quite long and error-prone to read/write.

Refs:
 - https://issues.redhat.com/browse/ACM-18626

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>